fix(realty-image): fix realty image domain
