+++ src/realty/domain/realty-image.exception.ts
